Files and namespace renames and corrections completed.
