+++ CoordConverterUI/CoordinateConverterCmd.csproj
