+++ CoordinateConversionUtility.sln
@@ -7,7 +7,7 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CoordinateConversionUtility
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CoordinateConversionUtility_UnitTests", "CoordinateConversionUtility_UnitTests\CoordinateConversionUtility_UnitTests.csproj", "{E5D80EB3-6536-42BF-9483-A8CC314C0E95}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CoordConverterUI", "CoordConverterUI\CoordConverterUI.csproj", "{1ECD4469-7156-49B1-9B77-F42130249DFC}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CoordinateConverterCmd", "CoordConverterUI\CoordinateConverterCmd.csproj", "{1ECD4469-7156-49B1-9B77-F42130249DFC}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{4E54B677-EF2E-426B-B390-2B24731A3542}"
 	ProjectSection(SolutionItems) = preProject
