+++ CoordinateConversionUtility/CoordinateConversionUtility.csproj
@@ -48,7 +48,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="CoordinateConversionUtil.cs" />
+    <Compile Include="GridDdmExpert.cs" />
     <Compile Include="Helpers\ConversionHelper.cs" />
     <Compile Include="Helpers\InputHelper.cs" />
     <Compile Include="Helpers\LookupTablesHelper.cs" />
